Remove song window added
When user right clicks on a song  context menu shows up with play and delete. If user clicks on play, the same event handler from the play button will be called. If user click remove, the remove song window for confirmation will pop up asking for confirmation. Still figuring out the logic to delete the song